requirements.txt
transformers==4.48.3
torch
torchvision
torchaudio
python-dotenv
docx2txt
pdfplumber
